refactor: break down WhitelistMethodsCard
